+++ SUI Koober/App/KooberDependencyContainer.swift
@@ -34,7 +34,21 @@ final class KooberDependencyContainer {
   /// Data store that holds the authenticated user's session. This uses a fake implementation to simulate whether a user is signed when the app launches.
   let userSessionStore: UserSessionStore = FakeUserSessionStore(userAlreadySignedIn: false)
   
-  func makeAnonymousUserDependencyContainer() -> AnonymousUserDependencyContainer {
-    return AnonymousUserDependencyContainer(userSessionStore: self.userSessionStore)
+  // MARK: Factories
+  /// Factory method for creating a new sign in use case to sign users into Koober.
+  func makeSignInUseCase(username: String, password: String) -> SignInUseCase {
+    // Gather dependencies.
+    let remoteAPI = makeUserAuthenticationRemoteAPI()
+    let userSessionStore = self.userSessionStore
+    // Make use case.
+    return SignInUseCase(username: username,
+                         password: password,
+                         remoteAPI: remoteAPI,
+                         userSessionStore: userSessionStore)
+  }
+  
+  /// Makes a new user authentication remote API for authenticating users using the cloud.
+  func makeUserAuthenticationRemoteAPI() -> UserAuthenticationRemoteAPI {
+    return FakeUserAuthenticationRemoteAPI()
   }
 }
