+++ SUI Koober/App/MainView.swift
@@ -29,7 +29,7 @@
 import SwiftUI
 
 /// The app's root view.
-struct KooberiOSApp : View {
+struct MainView : View {
   /// This is the app's state store that is threaded down the view hierachy. It's a kind of Redux-like usage of SwiftUI. Anytime any state in `Koober` changes, this view and the entire view hierarchy is recomputed to reflect changes. This is supposed to be cheap because only the views that change are re-rendered.
   @ObjectBinding var koober: Koober
   
@@ -42,7 +42,7 @@ struct KooberiOSApp : View {
     case .launching:
       return AnyView(LaunchingView())
     case .running(let userState):
-      return AnyView(RunningView(userState: userState))
+      return AnyView(RunningView(userState: userState, koober: koober))
     }
   }
 }
