+++ SUI Koober/App/Running/Onboard/SignInViewModel.swift
@@ -26,20 +26,40 @@
 /// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 /// THE SOFTWARE.
 
-import Foundation
 import SwiftUI
-import Combine
 
-class AuthenticatedKoober: BindableObject {
-  var didChange = PassthroughSubject<Void, Never>()
+protocol SignInViewModelProtocol {
   
-  private(set) var userState: UserState {
-    didSet {
-      didChange.send(())
-    }
+  var email: Binding<String> { get }
+  var password: Binding<String> { get }
+  
+  func signIn()
+}
+
+
+final class SignInViewModel {
+  
+  private var _email = ""
+  private var _password = ""
+  
+  private let startSignInUseCase: (String, String) -> Void
+  
+  init(startSignInUseCase: @escaping (String, String) -> Void) {
+    self.startSignInUseCase = startSignInUseCase
+  }
+}
+
+extension SignInViewModel: SignInViewModelProtocol {
+  
+  var email: Binding<String> {
+    return .init(getValue: { self._email }, setValue: { self._email = $0 })
   }
   
-  init(userState: UserState) {
-    self.userState = userState
+  var password: Binding<String> {
+    return .init(getValue: { self._password }, setValue: { self._password = $0 })
+  }
+  
+  func signIn() {
+    startSignInUseCase(_email, _password)
   }
 }
